split comparator changelog results in two components
