+++ src/components/RepositoryReleasesChangelog.tsx
@@ -1,80 +1,15 @@
-import {
-	Alert,
-	AlertIcon,
-	Box,
-	CircularProgress,
-	Flex,
-	Heading,
-	Skeleton,
-	Stack,
-} from '@chakra-ui/react'
+import { Alert, AlertIcon, CircularProgress, Flex } from '@chakra-ui/react'
 import { useEffect, useState } from 'react'
 
-import ProcessedReleaseChangeDescription from '~/components/ProcessedReleaseChangeDescription'
-import TextSkeleton from '~/components/TextSkeleton'
-import useProcessReleases from '~/hooks/useProcessReleases'
-import type {
-	ProcessedRelease,
-	Release,
-	ReleaseGroup,
-	ReleaseVersion,
-	Repository,
-} from '~/models'
+import ComparatorChangelogResults from '~/components/ComparatorChangelogResults'
+import type { Release, ReleaseVersion, Repository } from '~/models'
 import { useReleasesQuery } from '~/queries/release'
-import {
-	compareReleaseGroupsByPriority,
-	filterReleasesByVersionRange,
-	compareReleasesByVersion,
-} from '~/utils'
+import { compareReleasesByVersion, filterReleasesByVersionRange } from '~/utils'
 
 interface RepositoryReleasesChangelogProps {
 	repository: Repository
 	fromVersion?: ReleaseVersion
 	toVersion?: ReleaseVersion
-}
-
-const ReleaseChangelogGroup = ({
-	title,
-	releaseGroup,
-	repository,
-	shouldShowTitle,
-}: {
-	title: ReleaseGroup
-	releaseGroup: Array<ProcessedRelease>
-	repository: Repository
-	shouldShowTitle: boolean
-}) => {
-	const textTransform =
-		title === 'breaking changes' ? 'uppercase' : 'capitalize'
-
-	return (
-		<Box key={title}>
-			{shouldShowTitle && (
-				<Heading
-					as="h2"
-					size="xl"
-					bgColor="background3"
-					mb={4}
-					py={4}
-					textTransform={textTransform}
-					position="sticky"
-					top={0}
-				>
-					{title}
-				</Heading>
-			)}
-			<Box mb={4}>
-				{releaseGroup.map((processedReleaseChange: ProcessedRelease) => (
-					<ProcessedReleaseChangeDescription
-						key={`${title}-${processedReleaseChange.id}`}
-						repository={repository}
-						processedReleaseChange={processedReleaseChange}
-						mb={8}
-					/>
-				))}
-			</Box>
-		</Box>
-	)
 }
 
 const RepositoryReleasesChangelog = ({
@@ -84,9 +19,6 @@ const RepositoryReleasesChangelog = ({
 }: RepositoryReleasesChangelogProps) => {
 	const [filteredReleases, setFilteredReleases] =
 		useState<Array<Release> | null>(null)
-
-	const { processedReleases, isProcessing } =
-		useProcessReleases(filteredReleases)
 
 	const { data: releases, isFetching } = useReleasesQuery({
 		repository,
@@ -107,20 +39,6 @@ const RepositoryReleasesChangelog = ({
 		}
 	}, [fromVersion, releases, toVersion])
 
-	const shouldShowProcessedReleaseTitle = (() => {
-		if (!processedReleases) {
-			return false
-		}
-
-		const groupsTitles = Object.keys(processedReleases)
-
-		return groupsTitles.length > 1 || !groupsTitles.includes('others')
-	})()
-
-	const sortedGroupTitles: Array<string> | null = processedReleases
-		? Object.keys(processedReleases).sort(compareReleaseGroupsByPriority)
-		: []
-
 	return (
 		<>
 			{isFetching && (
@@ -133,37 +51,20 @@ const RepositoryReleasesChangelog = ({
 					/>
 				</Flex>
 			)}
-			{isProcessing && (
-				<>
-					<Skeleton width="20%" height={8} mb={4} />
-					<TextSkeleton />
-				</>
+
+			{!isFetching && filteredReleases && (
+				<ComparatorChangelogResults
+					releases={filteredReleases}
+					repository={repository}
+				/>
 			)}
 
-			{!isProcessing && !isFetching && processedReleases && (
-				<Stack spacing={6}>
-					{sortedGroupTitles.map((title) => (
-						<ReleaseChangelogGroup
-							key={title}
-							title={title}
-							releaseGroup={processedReleases[title]}
-							repository={repository}
-							shouldShowTitle={shouldShowProcessedReleaseTitle}
-						/>
-					))}
-				</Stack>
+			{fromVersion && toVersion && !filteredReleases && !isFetching && (
+				<Alert status="error">
+					<AlertIcon />
+					No processed releases to show
+				</Alert>
 			)}
-
-			{fromVersion &&
-				toVersion &&
-				!processedReleases &&
-				!isProcessing &&
-				!isFetching && (
-					<Alert status="error">
-						<AlertIcon />
-						No processed releases to show
-					</Alert>
-				)}
 
 			<style global jsx>{`
 				.hljs-comment,
